+++ src/tui/mod.rs
@@ -5,7 +5,7 @@ pub mod option_select;
 use crossterm::{cursor, terminal, ExecutableCommand};
 use std::io;
 
-pub fn refresh_display(lines: u16) {
+pub fn refresh_display(lines: i32) {
     for _ in 0..lines {
         io::stdout().execute(cursor::MoveUp(1)).unwrap();
         io::stdout()
